Move spec_to_pipeline to rechunk module
